requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.2
django-crispy-forms==2.3
gunicorn==23.0.0
packaging==24.1
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
python-dotenv==1.0.0  # or the latest version
